requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv
flask_admin==1.6.0
psycopg[binary]==3.1.11
SQLAlchemy==2.0.19
gunicorn==20.1.0
alembic==1.11.1